Fea: Added Shape stage
