+++ laravel/app/Domains/Aspect/database/seeders/Shape/ShapeFileSeeder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Domains\Aspect\database\seeders;
+namespace App\Domains\Aspect\database\seeders\Shape;
 
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
