--- laravel/app/Domains/Aspect/src/Models/Stages/ShapeCategory.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Aspect\Models\Stages;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class ShapeCategory extends Model
-{
-    use HasFactory;
-}
